Separated backend tests from static server ones
